# `itoa`: Can be about 10 times faster